Refactor: split WarningBlock.ts over multiple files, in a new directory
Also add functionality to restore a block that has been removed
